requirements.txt
streamlit
face_recognition
numpy
Pillow
dlib